fix: expose creator dashboard
